requirements.txt
# Core Dependencies
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0

# ML Models
openai-whisper==20250625
transformers==4.57.0
huggingface-hub==0.35.3
safetensors==0.6.2
tokenizers==0.22.1

# Vision Model Integration
ollama==0.6.0

# Audio Processing
librosa==0.11.0
soundfile==0.13.1
audioread==3.0.1
soxr==1.0.0
pydub==0.25.1

# Video Processing
opencv-python==4.12.0.88
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1

# YouTube Download
yt-dlp==2025.9.26

# Data Processing
numpy==2.2.6
pandas==2.3.3
scipy==1.16.2
scikit-learn==1.7.2
pillow==11.3.0

# Configuration & Utilities
PyYAML==6.0.3
python-dotenv==1.1.1
pydantic==2.11.10
pydantic_core==2.33.2
httpx==0.28.1
requests==2.32.5
tqdm==4.67.1
joblib==1.5.2

# Social scraping
snscrape==0.7.0.20230622

# CLI & Logging
rich==14.1.0
colorlog==6.9.0

# ML Acceleration (Apple Silicon)
accelerate==1.10.1

# System Utilities
filelock==3.19.1
fsspec==2025.9.0
packaging==25.0
platformdirs==4.4.0
psutil==7.1.0
